+++ lib/mixins/pushable.dart
@@ -18,13 +18,10 @@ mixin Pushable on ObjectCollision {
   bool enablePushable = true;
 
   @override
-  bool onCollision(GameComponent component, bool active) {
+  void onCollisionHappened(GameComponent component, bool active) {
     if (enablePushable) {
       if (this is Movement) {
-        if (!active && component is Movement) {
-          if (!onPush(component)) {
-            return super.onCollision(component, active);
-          }
+        if (!active && component is Movement && onPush(component)) {
           Vector2 displacement = center - component.center;
           if (displacement.x.abs() > displacement.y.abs()) {
             if (displacement.x < 0) {
@@ -46,7 +43,7 @@ mixin Pushable on ObjectCollision {
             'The mixin Pushable not working in ($this) because this component don`t have the `Movement` mixin');
       }
     }
-    return super.onCollision(component, active);
+    super.onCollisionHappened(component, isVisible);
   }
 
   /// Returning true if the component is pushable, false otherwise.