+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [2.12.3]
+- Fix issue [#379](https://github.com/RafaelBarbosatec/bonfire/issues/379)
+- Adds in `ObjectCollision` the method `onCollisionHappened`
+
 # [2.12.2]
 - Adds `FollowerObject`. Thanks [Matt El Mouktafi](https://github.com/mel-mouk)!